requirements.txt
google-genai
pypdf
pillow
streamlit
# Add others as you implement docx, audio/video processing